pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Артем Готин <aremgotin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.0"
pydantic-settings = "^2.4.0"
uvicorn = "^0.30.5"
isort = "^5.13.2"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.32"}
asyncpg = "^0.29.0"
pydantic = {extras = ["email"], version = "^2.8.2"}
fastapi-jwt = "^0.3.0"
python-multipart = "^0.0.9"
passlib = {version = "^1.7.4", extras = ["bcrypt"]}
fastapi-mail = "^1.4.1"
pytz = "^2024.1"
fastapi-jwt-auth = "^0.5.0"
httpx = "^0.27.2"
python-json-logger = "^2.0.7"
numpy = "^2.2.6"
docker = "^7.1.0"
python-docker = "^0.2.0"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
